Factorize skills tags
