+++ packages/web/components/Quest/QuestTags.tsx
@@ -1,13 +1,11 @@
-import { MetaTag, Tooltip, Wrap, WrapItem } from '@metafam/ds';
+import { MetaTag, Tooltip } from '@metafam/ds';
 import {
   QuestCompletionStatus_Enum,
   QuestRepetition_Enum,
   QuestStatus_Enum,
-  Skill,
 } from 'graphql/autogen/types';
 import React from 'react';
 
-import { SkillColors } from '../../graphql/types';
 import { QuestRepetitionHint } from '../../utils/questHelpers';
 
 export const RepetitionColors: Record<QuestRepetition_Enum, string> = {
@@ -64,34 +62,4 @@ export const CompletionStatusTag: React.FC<QuestCompletionProps> = ({
   >
     {status}
   </MetaTag>
-);
-
-interface SkillsProps {
-  skills: Skill[];
-  maxSkills?: number;
-}
-export const SkillsTags: React.FC<SkillsProps> = ({
-  maxSkills = 4,
-  skills,
-}) => (
-  <Wrap>
-    {skills.slice(0, maxSkills).map((skill) => (
-      <WrapItem key={skill.id}>
-        <MetaTag
-          size="md"
-          fontWeight="normal"
-          backgroundColor={SkillColors[skill.category]}
-        >
-          {skill.name}
-        </MetaTag>
-      </WrapItem>
-    ))}
-    {skills.length > maxSkills && (
-      <WrapItem>
-        <MetaTag size="md" fontWeight="normal">
-          {`+${skills.length - maxSkills}`}
-        </MetaTag>
-      </WrapItem>
-    )}
-  </Wrap>
 );
